refactor(memory_ipc): extract tests to sibling (#1266)
Partial: memory_ipc.rs 865 -> 786 LOC. Tests moved to tests_memory_ipc.rs (~80 LOC). Source split deferred.
